requirements.txt
youtube-dl
youtube-search-python
telepotpro
python-dotenv